Add support for targetting MinGW64
This is done by allowing XAudio2 dependency to be replaced by FAudio.
It replaces all XAudio2 calls with an equivalent one in FAudio.

A `CMakeLists.txt` is added for building with CMake.
Together with FAudio, they add support for building in MinGW64.

\#### Details

FAudio is an reimplementation of XAudio.
It can be found at <https://github.com/FNA-XNA/FAudio>.
It is also available as a package in MinGW64,
from <https://packages.msys2.org/package/mingw-w64-x86_64-FAudio>.
So by using FAudio instead,
this allows `pcaudiolib` to be compiled in MinGW64.

There is a caveat to this replacement in implementation.
FAudio has no dependency on `<windows.h>`,
relying on SDL2 for cross-platform work.
This means its implementation has to define its own identifiers
to replace those in `<windows.h>`,
in order to provide the same API.
In particular, identifiers from FAudio headers can clash
with `<windows.h>` if both are used.

Unfortunately, pcaudiolib directly uses `<windows.h>`
for some Windows API functions and macros to use those APIs
(COM library memory management and `FormatMessageA`).
An `#undef` is used to workaround this
-- this is considered unclean --
and should be seen as a temporary fix
until the COM library dependency can be removed from the FAudio version.