Fix a conversion warning in MSVC build
  * Add a pointer cast to a `FormatMessageA` argument
    to remove a conversion warning.

\#### Details

One of the parameters to `FormatMessageA` is `LPSTR lpBuffer`.
But when the `dwFlags` argument sets `FORMAT_MESSAGE_ALLOCATE_BUFFER`,
the `lpBuffer` should instead be a pointer to `LPTSTR`.
According to its documentation at
<https://docs.microsoft.com/windows/win32/api/winbase/nf-winbase-formatmessagea>,
the caller should cast the `LPTSTR*` to `LPTSTR`
before passing it to the function.
This recommendation is applied to remove a compiler warning.