Fix CMake build not searching for dependencies
CMake provides a way to create and export a package configuration
that other CMake users can use to link to `pcaudiolib`.
If a `pcaudiolib` build depends on `FAudio`,
then the exported configuration from `pcaudiolib`
should search for the package configuration exported by `FAudio`,
to ensure the depenency is automatically satisfied.

This search has been added.