Fix MSVC build not installing `.lib` for linking
The MSVC build built using CMake creates only the DLL by default.
This DLL is then installed when the package is installed.
However, as a CMake dependency target by other CMake users,
having just the DLL is not sufficient for linking.
The corresponding symbols `.lib` should also be provided.
There is a CMake property to create this `.lib` file.
and it will also be installed when the DLL is.

The property for generating the `.lib` is now toggled to create the file.