Fix audio not playing after writing to device
The function `xaudio2_object_write` is expected
to start playing the audio data given to it as soon as possible.
However, due to copy-and-paste error,
a wrong flag was used it in its implementation.
In particular, the function did not send the command
to the underlying FAudio implementation to read the buffer.

A proper flag is now used.