Add a search for SDL2 when building with FAudio
FAudio is a reimplementation of XAudio2 using SDL2.
So it has SDL2 as a dependency.
Such dependencies are usually searched for and linked to automatically
when using CMake packages of libraries built with CMake.
However, the CMake package of FAudio does not search for SDL2.
This means users of `pcaudiolib` would need to manually search for it.

The search is now done automatically
by the CMake package of `pcaudiolib`,
both during build time of `pcaudiolib`,
and also transitively for CMake users linking to `pcaudiolib`.